Fixed FDPF Wrote the FDNR method docs

+++ src/GridCal/Engine/Simulations/PTDF/ptdf_ts_driver.py
@@ -308,7 +308,8 @@ class PtdfTimeSeries(QThread):
         Cancel the simulation
         """
         self.__cancel__ = True
-        self.pool.terminate()
+        if self.pool is not None:
+            self.pool.terminate()
         self.progress_signal.emit(0.0)
         self.progress_text.emit('Cancelled!')
         self.done_signal.emit()
