+++ src/GridCal/Engine/Simulations/PowerFlow/time_series_driver.py
@@ -711,7 +711,8 @@ class TimeSeries(QThread):
         Cancel the simulation
         """
         self.__cancel__ = True
-        self.pool.terminate()
+        if self.pool is not None:
+            self.pool.terminate()
         self.progress_signal.emit(0.0)
         self.progress_text.emit('Cancelled!')
         self.done_signal.emit()
@@ -1019,7 +1020,8 @@ class SampledTimeSeries(QThread):
         Cancel the simulation
         """
         self.__cancel__ = True
-        self.pool.terminate()
+        if self.pool is not None:
+            self.pool.terminate()
         self.progress_signal.emit(0.0)
         self.progress_text.emit('Cancelled!')
         self.done_signal.emit()
